aggiornato tutto con mia versione locale

+++ logic/process.py
@@ -15,7 +15,7 @@ def wow_filter(smap):
     with warnings.catch_warnings():
         warnings.simplefilter("ignore")
         clean_data = np.nan_to_num(smap.data, nan=np.nanmedian(smap.data))
-        clean_map = sunpy.map.Map(clean_data, smap.meta)
+        clean_map = sunpy.map.Map(clean_data.astype('float'), smap.meta)
         wow_map = enhance.wow(
             clean_map,
             bilateral=1,
